+++ src/Rules.php
@@ -4,18 +4,19 @@ declare(strict_types=1);
 
 namespace Pdp;
 
+use SplFileObject;
 use SplTempFileObject;
 use Stringable;
 use function array_pop;
 use function explode;
 use function preg_match;
-use function strpos;
 use function substr;
 
 final class Rules implements PublicSuffixList
 {
     private const ICANN_DOMAINS = 'ICANN_DOMAINS';
     private const PRIVATE_DOMAINS = 'PRIVATE_DOMAINS';
+    private const UNKNOWN_DOMAINS = 'UNKNOWN_DOMAINS';
 
     private const REGEX_PSL_SECTION = ',^// ===(?<point>BEGIN|END) (?<type>ICANN|PRIVATE) DOMAINS===,';
     private const PSL_SECTION = [
@@ -42,7 +43,7 @@ final class Rules implements PublicSuffixList
      * @param null|resource $context
      *
      * @throws UnableToLoadResource         If the rules can not be loaded from the path
-     * @throws UnableToLoadPublicSuffixList If the rules contains in the resource are invalid
+     * @throws UnableToLoadPublicSuffixList If the rules contain in the resource are invalid
      */
     public static function fromPath(string $path, $context = null): self
     {
@@ -52,7 +53,7 @@ final class Rules implements PublicSuffixList
     /**
      * Returns a new instance from a string.
      *
-     * @throws UnableToLoadPublicSuffixList If the rules contains in the resource are invalid
+     * @throws UnableToLoadPublicSuffixList If the rules contain in the resource are invalid
      */
     public static function fromString(Stringable|string $content): self
     {
@@ -70,11 +71,11 @@ final class Rules implements PublicSuffixList
         $section = '';
         $file = new SplTempFileObject();
         $file->fwrite($content);
-        $file->setFlags(SplTempFileObject::DROP_NEW_LINE | SplTempFileObject::READ_AHEAD | SplTempFileObject::SKIP_EMPTY);
+        $file->setFlags(SplFileObject::DROP_NEW_LINE | SplFileObject::READ_AHEAD | SplFileObject::SKIP_EMPTY);
         /** @var string $line */
         foreach ($file as $line) {
             $section = self::getSection($section, $line);
-            if (in_array($section, [self::PRIVATE_DOMAINS, self::ICANN_DOMAINS], true) && false === strpos($line, '//')) {
+            if (in_array($section, [self::PRIVATE_DOMAINS, self::ICANN_DOMAINS], true) && !str_contains($line, '//')) {
                 $rules[$section] = self::addRule($rules[$section], explode('.', $line));
             }
         }
@@ -159,7 +160,7 @@ final class Rules implements PublicSuffixList
             return $this->getCookieDomain($host);
         } catch (UnableToResolveDomain $exception) {
             return ResolvedDomain::fromUnknown($exception->domain());
-        } catch (SyntaxError $exception) {
+        } catch (SyntaxError) {
             return ResolvedDomain::fromUnknown(Domain::fromIDNA2008(null));
         }
     }
@@ -170,7 +171,7 @@ final class Rules implements PublicSuffixList
     public function getCookieDomain($host): ResolvedDomainName
     {
         $domain = $this->validateDomain($host);
-        [$suffixLength, $section] = $this->resolveSuffix($domain, '');
+        [$suffixLength, $section] = $this->resolveSuffix($domain, self::UNKNOWN_DOMAINS);
 
         return match (true) {
             self::ICANN_DOMAINS === $section => ResolvedDomain::fromICANN($domain, $suffixLength),
@@ -231,15 +232,17 @@ final class Rules implements PublicSuffixList
     }
 
     /**
-     * Returns the length and the section of thhe resolved effective top level domain.
+     * Returns the length and the section of the resolved effective top level domain.
      *
-     * @return array{0: int, 1:string}
+     * @param Rules::UNKNOWN_DOMAINS|Rules::ICANN_DOMAINS|Rules::PRIVATE_DOMAINS $section
+     *
+     * @return array{0: int, 1:Rules::UNKNOWN_DOMAINS|Rules::ICANN_DOMAINS|Rules::PRIVATE_DOMAINS}
      */
     private function resolveSuffix(DomainName $domain, string $section): array
     {
         $icannSuffixLength = $this->getPublicSuffixLengthFromSection($domain, self::ICANN_DOMAINS);
         if (1 > $icannSuffixLength) {
-            return [1, ''];
+            return [1, self::UNKNOWN_DOMAINS];
         }
 
         if (self::ICANN_DOMAINS === $section) {