+++ src/TopLevelDomainsTest.php
@@ -6,14 +6,12 @@ namespace Pdp;
 
 use DateTimeImmutable;
 use DateTimeZone;
+use PHPUnit\Framework\Attributes\DataProvider;
 use PHPUnit\Framework\TestCase;
 use Stringable;
 use TypeError;
 use function dirname;
 
-/**
- * @coversDefaultClass \Pdp\TopLevelDomains
- */
 final class TopLevelDomainsTest extends TestCase
 {
     private static TopLevelDomains $topLevelDomains;
@@ -23,12 +21,6 @@ final class TopLevelDomainsTest extends TestCase
         self::$topLevelDomains = TopLevelDomains::fromPath(dirname(__DIR__).'/test_data/tlds-alpha-by-domain.txt');
     }
 
-    /**
-     * @covers ::fromPath
-     * @covers ::fromString
-     * @covers \Pdp\Stream
-     * @covers ::__construct
-     */
     public function testCreateFromPath(): void
     {
         $context = stream_context_create([
@@ -43,10 +35,6 @@ final class TopLevelDomainsTest extends TestCase
         self::assertEquals(self::$topLevelDomains, $topLevelDomains);
     }
 
-    /**
-     * @covers ::fromPath
-     * @covers \Pdp\UnableToLoadTopLevelDomainList
-     */
     public function testCreateFromPathThrowsException(): void
     {
         $this->expectException(UnableToLoadResource::class);
@@ -54,9 +42,7 @@ final class TopLevelDomainsTest extends TestCase
         TopLevelDomains::fromPath('/foo/bar.dat');
     }
 
-    /**
-     * @dataProvider invalidContentProvider
-     */
+    #[DataProvider('invalidContentProvider')]
     public function testConverterThrowsException(string $content): void
     {
         $this->expectException(UnableToLoadTopLevelDomainList::class);
@@ -67,7 +53,7 @@ final class TopLevelDomainsTest extends TestCase
     /**
      * @return iterable<string,array{0:string}>
      */
-    public function invalidContentProvider(): iterable
+    public static function invalidContentProvider(): iterable
     {
         $doubleHeader = <<<EOF
 # Version 2018082200, Last Updated Wed Aug 22 07:07:01 2018 UTC
@@ -128,10 +114,6 @@ EOF;
         ];
     }
 
-    /**
-     * @covers ::__set_state
-     * @covers ::__construct
-     */
     public function testSetState(): void
     {
         /** @var TopLevelDomains $topLevelDomains */
@@ -160,9 +142,7 @@ EOF;
         }
     }
 
-    /**
-     * @dataProvider validDomainProvider
-     */
+    #[DataProvider('validDomainProvider')]
     public function testResolve(DomainNameProvider|Host|Stringable|string|int|null $tld): void
     {
         self::assertSame(
@@ -171,9 +151,7 @@ EOF;
         );
     }
 
-    /**
-     * @dataProvider validDomainProvider
-     */
+    #[DataProvider('validDomainProvider')]
     public function testGetTopLevelDomain(DomainNameProvider|Host|Stringable|string|int|null $tld): void
     {
         self::assertSame(
@@ -185,7 +163,7 @@ EOF;
     /**
      * @return iterable<string,array<object|string>>
      */
-    public function validDomainProvider(): iterable
+    public static function validDomainProvider(): iterable
     {
         $resolvedDomain = ResolvedDomain::fromICANN(Domain::fromIDNA2008('www.example.com'), 1);
 
@@ -278,7 +256,7 @@ EOF;
     /**
      * @return iterable<string,array<string|object>>
      */
-    public function validTldProvider(): iterable
+    public static function validTldProvider(): iterable
     {
         return [
             'simple TLD' => ['COM'],
@@ -304,7 +282,7 @@ EOF;
     /**
      * @return iterable<string,array<object|string|null>>
      */
-    public function invalidTldProvider(): iterable
+    public static function invalidTldProvider(): iterable
     {
         return [
             'invalid TLD (1)' => ['COMM'],